requirements.txt
# Jade Fitness Hub - Requirements
streamlit>=1.28.0
pandas>=2.0.0
yt-dlp>=2023.10.13
supabase>=2.0.0
python-dotenv>=1.0.0
